# interval: fix BenchmarkBTreeIterPrev