--- a/pkg/kv/kvserver/concurrency/keylocks_interval_btree_test.go
+++ b/pkg/kv/kvserver/concurrency/keylocks_interval_btree_test.go
@@ -1063,7 +1063,7 @@ func BenchmarkBTreeIterPrev(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		if !it.Valid() {
-			it.First()
+			it.Last()
 		}
 		it.Prev()
 	}