Informs #122530.

This benchmark was starting the backwards iteration from the front of the tree, not the rear, so it was calling `First` on every iteration.

```
name              old time/op  new time/op  delta
BTreeIterPrev-10  12.2ns ± 1%   2.3ns ± 3%  -81.25%  (p=0.000 n=10+9)
```

Release note: None